requirements.txt
cachetools==4.2.1
Flask==1.1.2
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
slackclient==2.9.3
slackeventsapi==2.2.1
tabulate==0.8.9
tradingview-ta==3.2.2
websocket-client==0.58.0
